+++ apps/demo.lasereyes.build/CHANGELOG.md
@@ -1,5 +1,21 @@
 # demo.lasereyes.build
 
+## 0.0.0
+
+### Patch Changes
+
+- 2450f0b: Integrate Alkanes and add oylnet
+- Updated dependencies
+  - @omnisat/lasereyes@0.0.150
+
+## 0.0.0-next.4
+
+### Patch Changes
+
+- Integrate Alkanes and add oylnet
+- Updated dependencies
+  - @omnisat/lasereyes@0.0.150-next.0
+
 ## 0.0.0-next.3
 
 ### Patch Changes