+++ packages/lasereyes-core/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @omnisat/lasereyes-core
 
+## 0.0.72
+
+### Patch Changes
+
+- 2450f0b: Integrate Alkanes and add oylnet
+
+## 0.0.72-next.0
+
+### Patch Changes
+
+- bd6058e: test
+- Integrate Alkanes and add oylnet
+
 ## 0.0.69-next.5
 
 ### Patch Changes